requirements.txt
Django==3.1.6
tblib==1.7.0
psycopg2==2.8.6
Fabric3==1.14.post1
gunicorn==20.0.4
django-autoslug==1.9.8
